+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/cache/RedisBatchBaseCache.java
@@ -9,28 +9,33 @@ import com.dwarfeng.subgrade.stack.cache.BatchBaseCache;
 import com.dwarfeng.subgrade.stack.exception.CacheException;
 import org.springframework.data.redis.core.RedisTemplate;
 
-import javax.annotation.Nullable;
-import java.util.ArrayList;
-import java.util.List;
+import javax.annotation.Nonnull;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 
 /**
  * 使用 Redis 实现的 BaseCache。
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.0.1-beta
  */
 public class RedisBatchBaseCache<K extends Key, E extends Entity<K>, JE extends Bean> implements BatchBaseCache<K, E> {
 
+    @Nonnull
     private RedisTemplate<String, JE> template;
+    @Nonnull
     private StringKeyFormatter<K> formatter;
+    @Nonnull
     private BeanTransformer<E, JE> transformer;
 
     public RedisBatchBaseCache(
-            @Nullable RedisTemplate<String, JE> template,
-            @Nullable StringKeyFormatter<K> formatter,
-            @Nullable BeanTransformer<E, JE> transformer) {
+            @Nonnull RedisTemplate<String, JE> template,
+            @Nonnull StringKeyFormatter<K> formatter,
+            @Nonnull BeanTransformer<E, JE> transformer
+    ) {
         this.template = template;
         this.formatter = formatter;
         this.transformer = transformer;
@@ -75,14 +80,20 @@ public class RedisBatchBaseCache<K extends Key, E extends Entity<K>, JE extends 
     @Override
     public void clear() throws CacheException {
         try {
-            template.delete(template.keys(formatter.generalFormat()));
+            Set<String> keys = Optional.ofNullable(
+                    template.keys(formatter.generalFormat())
+            ).orElse(Collections.emptySet());
+            template.delete(keys);
         } catch (Exception e) {
             throw new CacheException(e);
         }
     }
 
     private boolean internalExists(K key) {
-        return template.hasKey(formatKey(key));
+        // 获得装箱结果。
+        Boolean result = template.hasKey(formatKey(key));
+        // 拆箱并返回。
+        return result != null && result;
     }
 
     private E internalGet(K key) {
@@ -165,27 +176,39 @@ public class RedisBatchBaseCache<K extends Key, E extends Entity<K>, JE extends 
         return formatter.format(key);
     }
 
+    @Nonnull
     public RedisTemplate<String, JE> getTemplate() {
         return template;
     }
 
-    public void setTemplate(@Nullable RedisTemplate<String, JE> template) {
+    public void setTemplate(@Nonnull RedisTemplate<String, JE> template) {
         this.template = template;
     }
 
+    @Nonnull
     public StringKeyFormatter<K> getFormatter() {
         return formatter;
     }
 
-    public void setFormatter(@Nullable StringKeyFormatter<K> formatter) {
+    public void setFormatter(@Nonnull StringKeyFormatter<K> formatter) {
         this.formatter = formatter;
     }
 
+    @Nonnull
     public BeanTransformer<E, JE> getTransformer() {
         return transformer;
     }
 
-    public void setTransformer(@Nullable BeanTransformer<E, JE> transformer) {
+    public void setTransformer(@Nonnull BeanTransformer<E, JE> transformer) {
         this.transformer = transformer;
+    }
+
+    @Override
+    public String toString() {
+        return "RedisBatchBaseCache{" +
+                "template=" + template +
+                ", formatter=" + formatter +
+                ", transformer=" + transformer +
+                '}';
     }
 }