+++ subgrade-impl/src/main/java/com/dwarfeng/subgrade/impl/cache/RedisKeyListCache.java
@@ -11,29 +11,34 @@ import com.dwarfeng.subgrade.stack.exception.CacheException;
 import org.springframework.data.redis.core.RedisTemplate;
 
 import javax.annotation.Nonnull;
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 
 /**
  * 使用 Redis 实现的 ListCache。
- * <p>该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。</p>
+ *
+ * <p>
+ * 该类只提供最基本的方法实现，没有添加任何事务，请通过代理的方式在代理类中添加事务。
  *
  * @author DwArFeng
  * @since 0.0.3-beta
  */
-public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, JE extends Bean> implements KeyListCache<K, E> {
+public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, JE extends Bean> implements
+        KeyListCache<K, E> {
 
+    @Nonnull
     private RedisTemplate<String, JE> template;
+    @Nonnull
     private StringKeyFormatter<K> formatter;
+    @Nonnull
     private BeanTransformer<E, JE> transformer;
 
     public RedisKeyListCache(
             @Nonnull RedisTemplate<String, JE> template,
             @Nonnull StringKeyFormatter<K> formatter,
-            @Nonnull BeanTransformer<E, JE> transformer) {
+            @Nonnull BeanTransformer<E, JE> transformer
+    ) {
         this.template = template;
         this.formatter = formatter;
         this.transformer = transformer;
@@ -42,7 +47,10 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
     @Override
     public boolean exists(K key) throws CacheException {
         try {
-            return template.hasKey(formatKey(key));
+            // 获得装箱后的结果。
+            Boolean result = template.hasKey(formatKey(key));
+            // 拆箱并返回。
+            return result != null && result;
         } catch (Exception e) {
             throw new CacheException(e);
         }
@@ -52,7 +60,10 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
     public int size(K key) throws CacheException {
         try {
             String formatKey = formatKey(key);
-            return template.opsForList().size(formatKey).intValue();
+            // 获得装箱后的结果。
+            Long result = template.opsForList().size(formatKey);
+            // 拆箱并返回。
+            return result == null ? 0 : result.intValue();
         } catch (Exception e) {
             throw new CacheException(e);
         }
@@ -64,10 +75,13 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
         try {
             String formatKey = formatKey(key);
             Long totalSize = template.opsForList().size(formatKey);
-            if (totalSize == 0) {
+            if (Objects.isNull(totalSize) || totalSize == 0) {
                 return new ArrayList<>();
             } else {
                 List<JE> range = template.opsForList().range(formatKey, 0, totalSize - 1);
+                if (Objects.isNull(range)) {
+                    return new ArrayList<>();
+                }
                 return range.stream().map(transformer::reverseTransform).collect(Collectors.toList());
             }
         } catch (Exception e) {
@@ -75,11 +89,15 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
         }
     }
 
+    @SuppressWarnings("DuplicatedCode")
     @Override
     public List<E> get(K key, int beginIndex, int maxEntity) throws CacheException {
         try {
             String formatKey = formatKey(key);
             List<JE> range = template.opsForList().range(formatKey, beginIndex, beginIndex + maxEntity - 1);
+            if (Objects.isNull(range)) {
+                return new ArrayList<>();
+            }
             return range.stream().map(transformer::reverseTransform).collect(Collectors.toList());
         } catch (Exception e) {
             throw new CacheException(e);
@@ -92,9 +110,15 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
         try {
             String formatKey = formatKey(key);
             Long totalSize = template.opsForList().size(formatKey);
+            if (Objects.isNull(totalSize) || totalSize == 0) {
+                return new ArrayList<>();
+            }
             long beginIndex = (long) pagingInfo.getRows() * pagingInfo.getPage();
             long endIndex = Math.min(totalSize, beginIndex + pagingInfo.getRows()) - 1;
             List<JE> range = template.opsForList().range(formatKey, beginIndex, endIndex);
+            if (Objects.isNull(range)) {
+                return new ArrayList<>();
+            }
             return range.stream().map(transformer::reverseTransform).collect(Collectors.toList());
         } catch (Exception e) {
             throw new CacheException(e);
@@ -159,7 +183,10 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
     @Override
     public void clear() throws CacheException {
         try {
-            template.delete(template.keys(formatter.generalFormat()));
+            Set<String> keys = Optional.ofNullable(
+                    template.keys(formatter.generalFormat())
+            ).orElse(Collections.emptySet());
+            template.delete(keys);
         } catch (Exception e) {
             throw new CacheException(e);
         }
@@ -169,6 +196,7 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
         return formatter.format(key);
     }
 
+    @Nonnull
     public RedisTemplate<String, JE> getTemplate() {
         return template;
     }
@@ -177,6 +205,7 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
         this.template = template;
     }
 
+    @Nonnull
     public StringKeyFormatter<K> getFormatter() {
         return formatter;
     }
@@ -185,11 +214,21 @@ public class RedisKeyListCache<K extends Key, E extends Entity<? extends Key>, J
         this.formatter = formatter;
     }
 
+    @Nonnull
     public BeanTransformer<E, JE> getTransformer() {
         return transformer;
     }
 
     public void setTransformer(@Nonnull BeanTransformer<E, JE> transformer) {
         this.transformer = transformer;
+    }
+
+    @Override
+    public String toString() {
+        return "RedisKeyListCache{" +
+                "template=" + template +
+                ", formatter=" + formatter +
+                ", transformer=" + transformer +
+                '}';
     }
 }